requirements.txt
flask
flask-socketio
eventlet
pybit
pandas
python-dotenv
plotly
